nous-core/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx>=0.28.1
redis==5.0.1
google-genai
google-api-python-client
edge-tts==6.1.8
yt-dlp>=2023.12.30
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
pillow==10.1.0
aiofiles==23.2.1
websockets>=13.0
pydantic>=2.5.0